wip: integrating new reader [skip ci]
